[Added] Uped buffs for entertainer / doctors to have the highest values in game only in the  Mos Eisley Cantina 150 Skill mod bonus for medics / docs / entertainers
Currently, all player city only offers 100 buff mind / medical rating this is an additional 50 bonus making the Mos Eisley Cantina the highest buffing structure in the game with a 150 medical and 150 mind bonus. The new player hub also has new terminals outside and all the trainers in the game. New characters freshly made will be automatically teleported to the new player hub/starting zone.

Buff bonuses are only given inside the cantina you must be in the active zone to apply the bonus to players.