Created: Senator Rank (Name subject to change)
This rank was created to be given to those on the server who have dedicated their time into helping new players, existing players, and over helping the server grow.
Notes: This is a visible rank in game, just like the EC, CSR, and Owner ranks. However this rank does not have any extra abilities. It is strictly for the visible title, so that others may know that this player has helped, and is continuing to help other players.